build: Don't reset the config back to default
Automatically modifying files that are under source control is not a
good idea. Any files that are automatically generated (or modified)
should not be under source control, and the build system should not
be dependent on their state at the beginning of the build. Therefore
there is no reason to 'reset' any state after a build is complete.

Signed-off-by: Joel Nider <joel@nider.org>