Do not allow secondary miss to go to prefetcher or add appropriate statistics
This miss information is already sent to the prefetcher in the past; as Champsim do not have MSHR structure that tracks down secondary misses for the same cache line, we should not train prefetchers on this miss. If we want to train prefetchers on this, we need to add the miss and access stats there. That is "sim_access" and "sim_miss".