Migrated documentation into docs directory (#386)
This patch regards my recent comment on #12.

The wiki has not lived up to expectations for documentation. The wiki is
only editable by maintainers and cannot be merged alongside the code
that it describes.

This patch uses Sphinx and GitHub Pages to host the documentation
instead. Each branch gets its documentation copied to the `gh-pages`
branch, where it is read by Pages.