Modify branch type deduction. (#372)
Some colleagues and I are working on an improved cvp2champsim converter
that better conveys some of the characteristics of the original CVP-1
traces to the ChampSim traces. A paper discussing the converter has been
accepted for publication at the IISWC’23 conference. Sorry if my
discussion is too verbose but I think it is important to understand the
changes we propose.

One of the things that was missing in the original converter is the
dependence of a branch from the previous instructions generating its
source registers. When converting a branch from the CVP-1 trace, the
original cvp2champsim sets the IP, SP, and flag registers as source
and/or destination registers adequately so that ChampSim correctly
identifies the branch type. In this process, however, the original
source registers present in the CVP-1 trace are not preserved. This
affects some conditional branches, indirect calls, and indirect jumps,
which have general-purpose source registers in the CVP-1 trace.
Consequently, the dependency between these branch instructions and
producer instructions is lost.

This is particularly important when a branch depends on a long-latency
load. In this case, if the dependency is kept, a branch misprediction
can only be corrected after the long latency load is resolved, exposing
the branch misprediction penalty on the critical path. In addition, the
negative impact on performance can be aggravated with a decoupled
front-end and frequent BTB misses, as the prefetching effect of the
decoupled front-end can be drastically reduced.

Our improved cvp2champsim converter preserves all the register
dependencies among instructions present in the CVP-1 trace. To this end,
in the case of branches, it adds the source registers present in the
CVP-1 trace instruction to the required (depending on the branch type)
IP, SP, and flag registers.

For conditional branches, we distinguish two cases. If the CVP-1 trace
instruction has a source register, we add the source register to the
converted instruction and avoid adding the flag register as the original
cvp2champsim does. Otherwise, if the CVP-1 trace instruction does not
have any source register, we keep adding the flag register, as done in
the original converter. Unfortunately, replacing the flag register with
the source register for some conditional branches, conflicts with the
way ChampSim identifies them.

The way we can address this shortcoming is by patching ChampSim to
change the current condition to identify a conditional branch. In
addition, we also need to modify the way ChampSim deduces indirect jumps
as their check is performed before checking if a branch is conditional
and we should avoid that conditional branches reading from others get
classified as indirect jumps.

The changes should be safe and keep compatibility with current traces.
In these traces, conditional branches read from the flag register and
thus they will still be identified as conditionals. Regarding indirect
branches, in x86 they are always absolute and thus should not read the
instruction pointer. We don't know traces for which the new branch type
deduction breaks backward compatibility.

Co-authored-by: Nathan Gober <ngober@protonmail.com>