Added strong types for bandwidths  (#400)
This patch adds a strong type that encapsulates the operation of
consuming a fixed number of operations, a very common operation in
ChampSim. Once initialized, the maximum bandwidth cannot be changed.
Instead, consuming the bandwidth reduces the amount available until it
is depleted. The class raises an exception if too much bandwidth is
consumed.