Respond to each merged request
CACHE and DRAM_CHANNEL merge requests and return a response for each
group of merged packets, saving their bandwidth.

In my understanding, it is unrealistic to save bandwidth this way
because a requester will need to search all requests in its queue and
update them at once. BOOMv3, an open-source processor, does not seem to
implement it either.

Return responses for all merged packets in a group and make sure all of
them consume the bandwidth.