Add Kanata logging
Kanata is a log format to record the behavior of processor pipelines,
which can be used with Konata, an instruction pipeline visualizer.

I'm planning to implement a feature to generate Kanata logs in a
gradual approach to minimize the burden for both me and the reviewers.
This is the first commit to add the basic infrastructure and instrument
to the instruction pipeline. Instrumenting the memory subsystem needs
significant code changes in the subsystem itself to track memory
accesses across merging and forwarding, which this commit does not
include.

However, the added Kanata logging infrastructure is already designed to
be versatile enough to cover the memory subsystem in the
future. More concretely:

- It has its own representation of "instruction"; currently it only
  logs ooo_model_instr, but it can also log prefetches in the future.
- It implements a concept of "lane", which is a unit of execution and
  part of an instruction. An instruction can have multiple lanes to
  track multiple memory accesses initiated by an instruction.
- A lane can keep its instruction alive even after the instruction
  retires from the ROB so that it can log the following stores, for
  example.

Please see the README.md change to know how to use this new feature.